+++ src/Tracker/PosLaju.php
@@ -16,7 +16,7 @@ class PosLaju extends BaseTracker
     protected $url = "https://poslaju.com.my/track-trace-v2/";
     protected $source = "Post Laju";
     protected $code = "poslaju";
-    protected $method = METHOD_POST;
+    protected $method = PARCEL_METHOD_POST;
 
     public function setTrackingNumber($refNum)
     {